# server/src/main/resources/openapi/family_accounting_openapi.yaml
...
openapi: 3.0.1
info:
  title: OpenAPI definition
  description: |
    家庭记账系统的 RESTful API，支持用户认证、人员管理、账户管理、交易记录和统计。
    支持银行、微信、支付宝等多种账号，每种账号下可有多张卡或子账户。
    角色权限：
    - 管理员：访问所有数据，管理用户、账号、子账户。
    - 普通用户：仅访问通过 user_sub_accounts 关联的子账户和自己的交易/统计。
  version: v0
servers:
  - url: 'http://localhost:8080'
    description: Generated server url
...
paths:
  '/users/{userId}':
    put:
      tags:
        - Users
        - users
      summary: 更新用户
      operationId: usersUpdate
      parameters:
        - name: userId
          in: path
          description: 用户 ID
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
        required: true
      responses:
        '200':
          description: 用户已更新
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '400':
          description: 无效输入（用户名重复或参数错误）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限不足（非管理员）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 用户不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
        - bearerAuth: []
    delete:
      tags:
        - Users
        - users
      summary: 删除用户
      operationId: usersDelete
      parameters:
        - name: userId
          in: path
          description: 用户 ID
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '204':
          description: 用户已删除
        '403':
          description: 权限不足（非管理员）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 用户不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 用户存在关联数据（如交易记录），无法删除
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
        - bearerAuth: []
  '/transactions/{id}':
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      tags:
        - Transactions
        - transactions
      summary: 获取交易记录详情
      operationId: getTransaction
      parameters: []
      responses:
        '200':
          description: 交易记录详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionDTO'
        '403':
          description: 权限不足（普通用户尝试访问未关联的数据）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 交易记录不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
        - bearerAuth: []
    put:
      tags:
        - Transactions
        - transactions
      summary: 更新交易记录
      operationId: updateTransaction
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionRequest'
        required: true
      responses:
        '200':
          description: 交易记录已更新
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionDTO'
        '400':
          description: 无效输入（交易类型不存在或参数格式错误）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限不足（普通用户尝试访问未关联的数据）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 交易记录不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
        - bearerAuth: []
    delete:
      tags:
        - Transactions
        - transactions
      summary: 删除交易记录
      operationId: deleteTransaction
      parameters: []
      responses:
        '204':
          description: 交易记录已删除
        '403':
          description: 权限不足（普通用户尝试访问未关联的数据）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 交易记录不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
        - bearerAuth: []
  '/transaction-types/{typeId}':
    put:
      tags:
        - transaction-types
        - TransactionTypes
      summary: 更新交易类型
      operationId: transactionTypesUpdate
      parameters:
        - name: typeId
          in: path
          description: 交易类型 ID
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionTypeRequest'
        required: true
      responses:
        '200':
          description: 交易类型已更新
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionTypeDTO'
        '400':
          description: 无效输入（名称重复或参数错误）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限不足（非管理员）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 交易类型不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
        - bearerAuth: []
    delete:
      tags:
        - transaction-types
        - TransactionTypes
      summary: 删除交易类型
      operationId: transactionTypesDelete
      parameters:
        - name: typeId
          in: path
          description: 交易类型 ID
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '204':
          description: 交易类型已删除
        '403':
          description: 权限不足（非管理员）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: 交易类型不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 交易类型存在关联交易记录，无法删除
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
        - bearerAuth: []
...
  /users:
    get:
      tags:
        - Users
        - users
      summary: 列出用户
      operationId: usersList
      responses:
        '200':
          description: 用户列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '403':
          description: 权限不足（非管理员）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
        - bearerAuth: []
    post:
      tags:
        - Users
        - users
      summary: 创建用户
      operationId: usersCreate
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
        required: true
      responses:
        '201':
          description: 用户已创建
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
        '400':
          description: 无效输入（用户名重复或参数错误）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限不足（非管理员）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
        - bearerAuth: []
...
  /transaction-types:
    get:
      tags:
        - transaction-types
        - TransactionTypes
      summary: 列出交易类型
      operationId: transactionTypesList
      responses:
        '200':
          description: 交易类型列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionTypeDTO'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
        - bearerAuth: []
    post:
      tags:
        - transaction-types
        - TransactionTypes
      summary: 创建交易类型
      operationId: transactionTypesCreate
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionTypeRequest'
        required: true
      responses:
        '201':
          description: 交易类型已创建
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionTypeDTO'
        '400':
          description: 无效输入（名称重复或参数错误）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: 权限不足（非管理员）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
        - bearerAuth: []
...
components:
  schemas:
    UserDTO:
      required:
        - role
        - userId
        - username
      type: object
      properties:
        userId:
          type: integer
          format: int32
        username:
          maxLength: 50
          minLength: 0
          type: string
        role:
          type: string
          enum:
            - ADMIN
            - USER
    ApiError:
      required:
        - code
        - message
      type: object
      properties:
        code:
          type: string
          enum:
            - PERMISSION_DENIED
            - DUPLICATE_ACCOUNT
            - INVALID_TRANSACTION
            - DATA_NOT_FOUND
            - CONFLICT
            - INTERNAL_ERROR
            - INVALID_REQUEST
            - VALIDATION_FAILED
            - INVALID_CREDENTIALS
        message:
          type: string
    UserRequest:
      required:
        - password
        - role
        - username
      type: object
      properties:
        username:
          maxLength: 50
          minLength: 0
          type: string
        password:
          maxLength: 100
          minLength: 0
          type: string
        role:
          type: string
          enum:
            - ADMIN
            - USER
    TransactionDTO:
      required:
        - amount
        - subAccountId
        - time
        - transactionId
        - typeId
        - userId
      type: object
      properties:
        transactionId:
          type: integer
          format: int32
        time:
          type: string
          format: date-time
        typeId:
          type: integer
          format: int32
        isIncome:
          type: boolean
          description: 是否为收入（由typeId关联的transaction_types.is_income自动生成）
        subAccountId:
          type: integer
          format: int32
        amount:
          pattern: '^\d+(\.\d{2})?$'
          type: string
        userId:
          type: integer
          format: int32
        remarks:
          maxLength: 255
          minLength: 0
          type: string
    TransactionRequest:
      required:
        - amount
        - subAccountId
        - time
        - typeId
      type: object
      properties:
        time:
          type: string
          format: date-time
        typeId:
          type: integer
          format: int32
        subAccountId:
          type: integer
          format: int32
        amount:
          pattern: '^\d+(\.\d{2})?$'
          type: string
        remarks:
          maxLength: 255
          minLength: 0
          type: string
    TransactionTypeDTO:
      required:
        - isIncome
        - typeId
        - typeName
      type: object
      properties:
        typeId:
          type: integer
          format: int32
        typeName:
          maxLength: 50
          minLength: 0
          type: string
        isIncome:
          type: boolean
    TransactionTypeRequest:
      required:
        - isIncome
        - typeName
      type: object
      properties:
        typeName:
          maxLength: 50
          minLength: 0
          type: string
          description: 交易类型名称
        isIncome:
          type: boolean
          description: 是否为收入，TRUE 为收入，FALSE 为支出
...